working on roles

+++ PokeGame.Orchestration/Handlers/Users/DeleteUserRequest.cs
@@ -2,13 +2,25 @@
 
 namespace PokeGame.Orchestration.Handlers.Users
 {
-    public class DeleteUserRequest :RequiredGuidRequest, IRequest { }
+    public class DeleteUserRequest : RequiredGuidRequest { }
 
     internal class DeleteUserHandler : DataHandler<DeleteUserRequest>
     {
         public DeleteUserHandler(IDataAccess dataAccess) : base(dataAccess) { }
 
-        public override async Task HandleRequestAsync(DeleteUserRequest request) =>
-            await _dataAccess.ExecuteAsync(new DeleteUser(request.Guid));
+        public override async Task HandleRequestAsync(DeleteUserRequest request)
+        {
+            if(!await _dataAccess.FetchAsync(new IsUserExistingWithGuid(request.Guid)))
+            {
+                throw new DoesNotExistException(typeof(User), request.Guid, nameof(request.Guid));
+            }
+
+            var rowsAffected = await _dataAccess.ExecuteAsync(new DeleteUser(request.Guid));
+
+            if (!rowsAffected.IsAnyRowsUpdated())
+            {
+                throw new ExpectationFailedException(nameof(DeleteUserRequest));
+            }
+        }
     }
 }
